requirements.txt
python-telegram-bot==22.1
httpx==0.28.1
pytest==8.2.1
pytest-asyncio==0.23.6
python-dotenv==1.1.0
requests==2.32.2
flask==3.0.2